Move PasswordHasher and related classes to security package
