+++ app/src/main/java/org/simple/clinic/security/PasswordHasher.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.login.applock
+package org.simple.clinic.security
 
 import io.reactivex.Single
