+++ app/src/test/java/org/simple/clinic/pin/PinEntryCardControllerTest.kt
@@ -11,8 +11,8 @@ import io.reactivex.Single
 import io.reactivex.subjects.PublishSubject
 import org.junit.Before
 import org.junit.Test
-import org.simple.clinic.login.applock.ComparisonResult
-import org.simple.clinic.login.applock.PasswordHasher
+import org.simple.clinic.security.ComparisonResult
+import org.simple.clinic.security.PasswordHasher
 import org.simple.clinic.patient.PatientMocker
 import org.simple.clinic.user.UserSession
 import org.simple.clinic.widgets.UiEvent
