+++ app/src/test/java/org/simple/clinic/user/UserSessionTest.kt
@@ -38,7 +38,7 @@ import org.simple.clinic.login.LoginApiV1
 import org.simple.clinic.login.LoginOtpSmsListener
 import org.simple.clinic.login.LoginResponse
 import org.simple.clinic.login.LoginResult
-import org.simple.clinic.login.applock.PasswordHasher
+import org.simple.clinic.security.PasswordHasher
 import org.simple.clinic.patient.PatientMocker
 import org.simple.clinic.patient.PatientRepository
 import org.simple.clinic.registration.FindUserResult
